requirements.txt
# requirements.txt
discord.py
python-dotenv
asqlite
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib